requirements.txt
# CAMEL Framework
camel-ai[all]

# Core Dependencies
openai>=1.0.0
anthropic>=0.40.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# NLP and Text Processing
rouge-score>=0.1.2
nltk>=3.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

